+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - New #322: Add `ColumnDefinitionBuilder` class (@Tigrov)
 - Enh #323: Refactor `Dsn` class (@Tigrov)
 - Enh #324: Set more specific result type in `Connection` methods `createCommand()` and `createTransaction()` (@vjik)
+- Enh #326: Refactor `Schema::normalizeDefaultValue()` method and move it to `ColumnFactory` class (@Tigrov)
 
 ## 1.2.0 March 21, 2024